cleanup and fixes

+++ hecl/extern/blowfish/blowfish.h
@@ -5,8 +5,8 @@ extern "C" {
 #include <stdlib.h>
 #include <stdint.h>
 
-void Blowfish_encipher(unsigned long *xl, unsigned long *xr);
-void Blowfish_decipher(unsigned long *xl, unsigned long *xr);
+void Blowfish_encipher(uint32_t *xl, uint32_t *xr);
+void Blowfish_decipher(uint32_t *xl, uint32_t *xr);
 int64_t Blowfish_hash(const void* buf, size_t len);
 
 #ifdef __cplusplus
